Fix Kenku name generation to be more in-line with lore
Also, this fixes changes on how the kenku names were utilised. It is more fantasy to have a metal kenku name for a fighter, but it really doesn't matter for NPC generation of this scale. It is also more realistic.

This allows Kenku's to associate themselves with multiple sounds as names.